curate: Validate records with a wrapper
This is necessary to register subparsers recursively.
