+++ augur/curate/apply_geolocation_rules.py
@@ -7,7 +7,7 @@ from augur.errors import AugurError
 from augur.io.print import print_err
 from augur.utils import first_line
 from augur.version import __version__
-from ._shared import shared_parser
+from ._shared import shared_parser, validate
 
 
 class NoGeolocationRulesProvidedError(AugurError):
@@ -231,6 +231,7 @@ def register_parser(parent_subparsers):
     return parser
 
 
+@validate
 def run(args, records):
 
     if args.no_default_rules and not args.geolocation_rules:
