+++ augur/curate/parse_genbank_location.py
@@ -9,7 +9,7 @@ import argparse
 from typing import Generator, List
 from augur.io.print import print_err
 from augur.utils import first_line
-from ._shared import shared_parser
+from ._shared import shared_parser, validate
 
 
 def parse_location(
