+++ augur/curate/passthru.py
@@ -3,7 +3,7 @@ Pass through records without doing any data transformations.
 Useful for testing, troubleshooting, or just converting file formats.
 """
 from augur.utils import first_line
-from ._shared import shared_parser
+from ._shared import shared_parser, validate
 
 
 def register_parser(parent_subparsers):
@@ -12,5 +12,6 @@ def register_parser(parent_subparsers):
         help=first_line(__doc__))
 
 
+@validate
 def run(args, records):
     yield from records
