+++ augur/curate/rename.py
@@ -8,7 +8,7 @@ from augur.argparse_ import ExtendOverwriteDefault
 from augur.io.print import print_err
 from augur.errors import AugurError
 from augur.utils import first_line
-from ._shared import shared_parser
+from ._shared import shared_parser, validate
 
 def register_parser(parent_subparsers):
     parser = parent_subparsers.add_parser("rename",
@@ -89,6 +89,7 @@ def transform_columns(existing_fields: List[str], field_map: List[Tuple[str,str]
     return m
 
 
+@validate
 def run(args: argparse.Namespace, records: Iterable[dict]) -> Iterable[dict]:
     col_map: Union[Literal[False], List[Tuple[str,str]]] = False
     for record in records:
