+++ augur/curate/titlecase.py
@@ -8,7 +8,7 @@ from augur.argparse_ import ExtendOverwriteDefault
 from augur.errors import AugurError
 from augur.io.print import print_err
 from augur.types import DataErrorMethod
-from ._shared import shared_parser
+from ._shared import shared_parser, validate
 
 def register_parser(parent_subparsers):
     parser = parent_subparsers.add_parser("titlecase",
@@ -75,6 +75,7 @@ def titlecase(text: Union[str, None], articles: Set[str] = set(), abbreviations:
     return ''.join(changecase(i, w) for i, w in words)
 
 
+@validate
 def run(args, records):
     failures = []
     failure_reporting = args.failure_reporting
